Theme
The Theme program has evolved over nearly 20 years (from
a Fortran IV version to a Delphi version) and now consists of
nearly one hundred thousand lines of source code. Therefore,
no source code is included in this paper, but a free copy of the
executable program for Windows 95/98/NT can be down-
loaded (www.hi.is/r-msm). In addition, at the time of this writ-
ing, a more powerful commercial version is being developed
and will be announced at the above site.